refactor: 랜덤 리스트 관련 Generator 추가
