+++ src/main/java/lotto/domain/Lottos.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.generator.LottoNumberGenerator;
+import lotto.generator.RandomLottNumberGenerator;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,11 +16,12 @@ public class Lottos {
 
     public static Lottos from(LottoBuyInfo buyInfo, ManualLottos manualLottos) {
         List<Lotto> newLottos = new ArrayList<>();
+        LottoNumberGenerator lottoNumberGenerator = new RandomLottNumberGenerator();
         for (int i = 0; i < manualLottos.size(); i++) {
             newLottos.add(Lotto.from(manualLottos.getManualLotto(i)));
         }
         for (int i = 0; i < buyInfo.getAutoCount(); i++) {
-            newLottos.add(Lotto.from(LottoNumbers.randomList()));
+            newLottos.add(Lotto.from(lottoNumberGenerator.getLottoNumberList()));
         }
         return new Lottos(newLottos);
     }
