+++ src/main/java/lotto/generator/LottoNumberGenerator.java
@@ -1,0 +1,9 @@
+package lotto.generator;
+
+import lotto.domain.LottoNumber;
+
+import java.util.List;
+
+public interface LottoNumberGenerator {
+    List<LottoNumber> getLottoNumberList();
+}
